feat: add 404 and init routing
