+++ package.json
@@ -14,8 +14,11 @@
     "coverage": "vitest run --coverage"
   },
   "dependencies": {
+    "@types/react-router-dom": "^5.3.3",
     "react": "^19.1.0",
-    "react-dom": "^19.1.0"
+    "react-dom": "^19.1.0",
+    "react-router": "^7.7.1",
+    "react-router-dom": "^7.7.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.30.1",
@@ -40,7 +43,7 @@
     "prettier": "3.6.2",
     "typescript": "~5.8.3",
     "typescript-eslint": "^8.35.1",
-    "vite": "^7.0.3",
+    "vite": "^7.0.6",
     "vitest": "^3.2.4"
   }
 }
