+++ vite.config.ts
@@ -13,7 +13,7 @@ export default defineConfig({
       provider: "v8",
       ignoreEmptyLines: true,
       thresholds: {
-        statements: 80,
+        statements: 70,
         branches: 50,
         functions: 50,
         lines: 50,
